Assign scorebet-bot to dependabot PRs (#13)
When dependabot PRs are created, this action fails because it attempts to assign `dependabot[bot]` to the generated PRs. That's not possible since `dependabot[bot]` is an app and not an actual bot account. This PR assigns a default `scorebet-bot` to dependabot PRs, but uses the `context.actor` (i.e. PR creator) for other PRs.